+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * Acrid model scale
   * Blind Pest feet rotation
   * Stone Golem model scale and particles
+  * Lemurian model scale and upper arm position
 
 **1.0.0 Changes:**